#14 list push_back fix & eps cmp add

+++ antsor/lab_01/lineseg.cpp
@@ -28,7 +28,7 @@ Point *LineSeg::p2()
 	return &end2;
 }
 
-float LineSeg::length()
+double LineSeg::length()
 {
 	return dist(end1, end2);
 }
